Create registry
