+++ david/__main__.py
@@ -1,10 +1,10 @@
 from flask import Flask, jsonify, make_response, request
 from flask_cors import CORS
 
-from david.adapters.adapter import AdaptEngine
 from david.assistant import Assistant
 from david.dialog import fetch_dialog
 from david.googleadap import GoogleWebHook
+from david.registry import Registry
 
 # from david.brain import fetch_model, fetch_know
 
@@ -14,7 +14,7 @@ CORS(app)
 assistant = Assistant()
 googleWH = GoogleWebHook(assistant)
 
-adapt_engine = AdaptEngine()
+registry = Registry.get_instance()
 
 
 @app.route("/")
@@ -33,7 +33,7 @@ def dialog():
     requestData = request.get_json()
 
     # [TODO] receive adapter from query string
-    adapter = adapt_engine.getAdapter()
+    adapter = registry.getAdapter()
 
     if not adapter.validade_data(requestData):
         # [TODO] send status code 400
