+++ david/adapters/adapter.py
@@ -34,28 +34,3 @@ class MessageAdapter(Adapter):
 
     def output(self, message: Message) -> Dict:
         return message.__dict__
-
-
-# [TODO] create a component registry and migrate this code
-class AdaptEngine:
-    adapters = {}
-
-    def __init__(self, defaultAdapter=None):
-
-        messageAdapter = MessageAdapter()
-        self.registryAdapter(messageAdapter)
-
-        if defaultAdapter:
-            self.defaultAdapter = defaultAdapter
-        else:
-            self.defaultAdapter = messageAdapter.name
-
-    def registryAdapter(self, adapter: Adapter):
-        self.adapters[adapter.name] = adapter
-
-    def getAdapter(self, adapterName=None):
-
-        if not adapterName:
-            adapterName = self.defaultAdapter
-
-        return self.adapters[adapterName]
